requirements.txt
pandas>=2.2
numpy>=1.26
matplotlib>=3.8
seaborn>=0.13
scikit-learn>=1.5
scipy>=1.12
statsmodels>=0.14
joblib>=1.4
python-docx>=1.1
